package/tbb: always build Release
When tbb is built in Debug mode, it installs libtbb_debug.so instead of
libtbb.so. This confuses downstream packages that want to link with it
(e.g. sysdig).

Always build in Release mode. This means that BR2_ENABLE_RUNTIME_DEBUG
has no effect for this package, but that shouldn't be a bi issue.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>