package/sysdig: fix DRIVER_NAME definition
When I wrote the previous commit, I was not fully focused and wrote:
-DDRIVER_NAME=$(FALCOSECURITY_LIBS_DRIVER_NAME.)
Which leaded to DRIVER_NAME to be empty.
So, it was not possible to use sysdig due to the following error message:
error opening device /dev/0. Make sure you have root credentials and that the  module is loaded.

Fixes: ea86757e51cf ("package/sysdig: bump to 0.29.1")
Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>