package/openocd: Fix segfault when using ST-Link driver
Recent changes in libusb have exposed a bug in OpenOCD which now crash when trying
to use the ST-Link driver.

Upstream has a fix as commit cff0e417da58adef1ceef9a63a99412c2cc87ff3. This add the commit
as a stand alone patch. The crash also happen on Linux, which was not mentionned in that
commit message.

Should be removed when OpenOCD is updated to a release newer than 0.11

Signed-off-by: Yannick Brosseau <yannick.brosseau@gmail.com>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
(cherry picked from commit 4239958963fd7da78fa8683646ad4b781a54c7fd)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>