package/libgcrypt: Fix powerpc64 build w/o optimisation
The following build failure was reported by the autobuilder:

powerpc64   |        libgcrypt-1.10.1        | NOK |
http://autobuild.buildroot.net/results/eb60bb95411076b92f5c63f9b21c6e576998031c | ORPH

This config is building with BR2_OPTIMIZE_0=y.

The code relies on the compiler turning the series of preprocessor
macros in to a constant for the second argument to vec_sld:

 chacha20-ppc.c:60:10: error: argument 3 must be a 4-bit unsigned literal
    60 |   return vec_sld (v, v, (16 - (4 * idx)) & 15);

The argument is a constant, but it would require a significant rework to
make this build without optimisation. Instead disable the vectorised
powerpc code when building with optimisation disabled.

Signed-off-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>