package/python-s3transfer: bump to version 0.6.0
While at it, remove the md5 hash as there is already a sha256 hash.

Note that the checksum of the license changed, but the license is
still the same (the only change in the license file is the removal of
a blank line at the end of the file).

Changelog: https://github.com/boto/s3transfer/blob/develop/CHANGELOG.rst

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>