package/gcnano-binaries: bump version to 6.4.7
Update to the latest version of the gcnano-binaries blobs, which are
compatible with the soon-to-be-released 5.15 kernel from the
STMicroelectronics BSP.

As we are installing all the libraries and the pkgconfig files we do not
need to list them anymore. Therefore this patch removes
GCNANO_BINARIES_LIBRARIES and GCNANO_BINARIES_PKG_CONFIGS variables.
It also updates the installation command to follow the path and naming
of the new version and the pkgconfig files to fill the prefix variable with
the right value.

The new version of gcnano has gbm.h header installed in a gbm folder,
this break the build of other packages. The patch fixes it with a simple
link.

Signed-off-by: Kory Maincent <kory.maincent@bootlin.com>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>