opendingux: Set default compressor to lzo-rle and load zram module
Allow the swap pages compressor to be user-selected, by setting the
SWAP_COMPRESSOR variable in /usr/local/etc/swap.conf. The default is now
the lzo-rle compressor.

As zram can now be built as a module, the init script will additionally
try to load the zram module before configuring swap.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>