opendingux: Update create-modules-fs.sh script
Instead of installing the kernel modules to a different directory in a
post-build script, package the modules installed to the target
directories into the modules.squashfs, remove them, and finally remove
the stamp file. This ensures that the kernel modules will be
re-installed into the target directories on the next build.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>