package/{python-}protobuf: bump to version 21.4
protobuf changed it versioning scheme [1]. The Release version is now only the
Minor.Patch version of the language specific version.

C++ stays on Major version 3, but python got bumped to 4. So the handling in
the .mk files changed to reflect that.

python-protobuf changed it version to 4 because of some breaking changes,
see[1].

[1] https://developers.google.com/protocol-buffers/docs/news/2022-05-06

Signed-off-by: Michael Nosthoff <buildroot@heine.tech>
Tested-by: Julien Olivain <ju.o@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>