package/pistache: fix dependency on rapidjson
rapidjson has been added as a dependency of pistache as part of commit
2f3218567953185008ec862f814bebecb9d6ed82 ("package/pistache: bump
version & switch to meson build"), using a "depends on". However, as
it is a library dependency, a "select" is normally used in Buildroot.

rapidjson's only dependency is on C++ support, which is already a
dependency of pistache.

In addition, use = instead of += for the rapidjson dependency in the
.mk file, as it's an unconditional dependency.

Cc: Thomas Ruschival <thomas@ruschival.de>
Cc: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
(cherry picked from commit 240aae4793022e4fe26ed7fa745d30e33a02fc4b)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>