package/lshw: bump to latest HEAD
lshw is seldom released, the last one being more than two years ago.
Although the amount of changes is not huge, there have been some
significant changes and fixes, so it warrants a bump to the current
HEAD of the git repo.

lshw tries to build the translations with msgfmt, but as a development
and debugging aid, or as a backend to other scripts, translations are
not really useful. We disable that by overriding the LANGUAGES
variables to an empty list.

Patches are dropped: the basename() patch has been superseded by an
upstream change (introduced shortname() as a wrapper to it), and the
LONG_BIT fix has been applied. It was a bit difficult to find, as
upstream has not applied the patches, nor kept authorship, nor kept the
commit logs...

Of noteworthy attention, this fixes the JSON output format (at least for
the flaws we were hitting in our cases).

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>