package/linux-tools: unbreak indentation for perf sub-options
Sub-options of a kconfig symbol are only preperly indented below that
sybol, if there is no inter-mixed sylbols or comments.

In this case, the comment about perf being unavailable is in-between the
perf symbol and the sub-options, which means the perf options are int
indented.

Move the comment before the definition of perf, so that the options
directly follow the per fsymbol, and so they are properly indented.

Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
(cherry picked from commit c475f3bc36a263bc7d3e1cdcd86fca0d48fbd27f)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>