package/network-manager: fix detection of pppd
With the switch to meson in 58c58466319a (package/network-manager:
migrate autotools build system to meson), the detection of pppd is
broken for cross-compilation:

    Has header "pppd/pppd.h" : YES
    Program pppd /sbin/pppd /usr/sbin/pppd found: NO

    ../../../O/br-2022.08/vtc/build/network-manager-1.36.4/meson.build:570:4: ERROR: Assert failed: pppd required but not found, please provide a valid pppd path or use -Dppp=false to disable it

Indeed, meson will simply look for existing files on the host, and that
can't work in cross-compilation.

Fix that by telling meson the results it can't find by itself.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>
Cc: James Hilliard <james.hilliard1@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>