package/go: security bump to version 1.17.13
go1.17.12 includes security fixes to the compress/gzip, encoding/gob,
encoding/xml, go/parser, io/fs, net/http, and path/filepath packages, as
well as bug fixes to the compiler, the go command, the runtime, and the
runtime/metrics package.

go1.17.13 includes security fixes to the encoding/gob and math/big packages,
as well as bug fixes to the compiler and the runtime.

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>