package/wpe/wpewebkit: add wpewebkit 2.38 variant
Introduce the BR2_PACKAGE_WPEWEBKIT2_38 symbol, used to choose a version
of WPE WebKit in the 2.38.x series. Official tarballs are used for the
sources.

A new symbol BR2_PACKAGE_WPEWEBKIT2_38_GST_DEPS centralizes all the
GStreamer dependencies imported from the 2.38.0 upstream recipe. While
some of these are most certainly missing when building 2.22 or 2.28,
they are only applied to 2.38 builds in order to keep the existing
behavior for older versions.

Signed-off-by: Adrian Perez de Castro <aperez@igalia.com>