package/wpe/wpebackend: add wpewebkit 2.38 variant
Add a BR2_PACKAGE_WPEBACKEND_LIBWPE_VERSION symbol to the Config
fragment which contains a string value with the libwpe to be used
with the chosen wpewebkit version. This contains the "1.14.0" value
when building wpewebkit 2.38.x, and "1.6.0" (the existing version)
for 2.28.x. Using a Config symbol avoids piling up conditionals in
the Make fragment, which is easier to maintain.

Signed-off-by: Adrian Perez de Castro <aperez@igalia.com>