package/wpe/wpewebkit: use ninja for compilation
Instruct CMake to use the Ninja generator, and override build/install
commands to use Ninja instead. This does not only result in faster
builds, it also allows building WPE WebKit 2.38.x without requiring
a CMake update due to usage of newer features which were implemented
in its Make generator in CMake 3.20. The host-ninja package is required
anyway to compile dependencies which use the Meson build system (like
libepoxy) so in practice this is not introducing a new host dependency
which would slow down the build.

Signed-off-by: Adrian Perez de Castro <aperez@igalia.com>